requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-picklefield==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.0.0
pipreqs==0.4.11
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
yarg==0.1.9
django-nested-admin==3.4.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
fpdf2==2.5.7
django-debug-toolbar==3.6.0
#idna==3.3
